Move enum tests to StringParserTests
